babel: Port integration tests to their own package
